requirements_new.txt
# AI Stock Chart Assistant - Production Dependencies
# Core GUI Framework
customtkinter>=5.2.0

# AI/ML APIs
google-generativeai>=0.3.0

# HTTP and networking
requests>=2.31.0
urllib3>=2.0.0

# Image processing and computer vision
Pillow>=10.0.0
opencv-python>=4.8.0

# Error handling and retry mechanisms
tenacity>=8.2.0

# Logging and monitoring
structlog>=23.1.0

# Environment and configuration
python-dotenv>=1.0.0

# Data handling and validation
pydantic>=2.0.0

# Export functionality
reportlab>=4.0.0

# Development and testing (optional)
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0

# Additional utilities
typing-extensions>=4.7.0
pathlib2>=2.3.0; python_version < "3.4"

# Additional dependencies
tkinter-tooltip==2.1.0
colorlog==6.8.0
watchdog==3.0.0
psutil==5.9.6
matplotlib==3.8.2
asyncio-throttle==1.0.2
concurrent-futures==3.1.1
configparser==6.0.0 